# Use a blank indentifier instead of an anonymous field for SDKShapeTraits