--- a/private/protocol/rest/payload.go
+++ b/private/protocol/rest/payload.go
@@ -12,7 +12,7 @@ func PayloadMember(i interface{}) interface{} {
 	if !v.IsValid() {
 		return nil
 	}
-	if field, ok := v.Type().FieldByName("SDKShapeTraits"); ok {
+	if field, ok := v.Type().FieldByName("_"); ok {
 		if payloadName := field.Tag.Get("payload"); payloadName != "" {
 			field, _ := v.Type().FieldByName(payloadName)
 			if field.Tag.Get("type") != "structure" {
@@ -34,7 +34,7 @@ func PayloadType(i interface{}) string {
 	if !v.IsValid() {
 		return ""
 	}
-	if field, ok := v.Type().FieldByName("SDKShapeTraits"); ok {
+	if field, ok := v.Type().FieldByName("_"); ok {
 		if payloadName := field.Tag.Get("payload"); payloadName != "" {
 			if member, ok := v.Type().FieldByName(payloadName); ok {
 				return member.Tag.Get("type")