… aws/corehandlers/param_validator_test.go
… private/protocol/ec2query/build_test.go
… private/protocol/ec2query/unmarshal_test.go
… private/protocol/jsonrpc/build_test.go
… private/protocol/jsonrpc/unmarshal_test.go
… private/protocol/query/build_test.go
… private/protocol/query/unmarshal_test.go
… private/protocol/restjson/build_test.go
… private/protocol/restjson/unmarshal_test.go
… private/protocol/restxml/build_test.go
… private/protocol/restxml/unmarshal_test.go
… service/apigateway/api.go
… service/autoscaling/api.go
… service/cloudfront/api.go
… service/cloudhsm/api.go
… service/cloudsearch/api.go
… service/cloudtrail/api.go
… service/cloudwatch/api.go
… service/cloudwatchlogs/api.go
… service/codecommit/api.go
… service/codedeploy/api.go
… service/codepipeline/api.go
… service/cognitoidentity/api.go
… service/cognitosync/api.go
… service/configservice/api.go
… service/datapipeline/api.go
… service/devicefarm/api.go
… service/directconnect/api.go
… service/directoryservice/api.go
… service/dynamodb/api.go