As far as I can tell this retains the current behavior with the following advantages:

1. The zero-width field has no overhead on instances of the struct. I think SDKShapeTraits is costing us up to 8 bytes on each struct instance (depending on packing/padding).
1. Cheaper to fetch via reflection. Benchmarks below.
1. More succinct.

One downside is that `_` is less descriptive than `SDKShapeTraits` but since it's an internal detail that doesn't seem like too big of a deal.

As mentioned in https://github.com/aws/aws-sdk-go/issues/377#issuecomment-157496324

## Example

### Old Style

```go
type InvokeAsyncInput struct {
  // The Lambda function name.
  FunctionName *string `location:"uri" locationName:"FunctionName" min:"1" type:"string" required:"true"`

  // JSON that you want to provide to your Lambda function as input.
  InvokeArgs io.ReadSeeker `type:"blob" required:"true"`

  metadataInvokeAsyncInput `json:"-" xml:"-"`
}

type metadataInvokeAsyncInput struct {
  SDKShapeTraits bool `type:"structure" payload:"InvokeArgs"`
}
```

### New Style
```go
type InvokeAsyncInput struct {
  _ struct{} `type:"structure" payload:"InvokeArgs"`

  // The Lambda function name.
  FunctionName *string `location:"uri" locationName:"FunctionName" min:"1" type:"string" required:"true"`

  // JSON that you want to provide to your Lambda function as input.
  InvokeArgs io.ReadSeeker `type:"blob" required:"true"`
}
```

# Protocol Benchmarks (vs master)

```
benchmark                                                                      old ns/op     new ns/op     delta
BenchmarkEC2QueryBuild_Complex_ec2AuthorizeSecurityGroupEgress-4               48412         48456         +0.09%
BenchmarkEC2QueryBuild_Simple_ec2AttachNetworkInterface-4                      12853         12925         +0.56%
BenchmarkBuildJSON-4                                                           7597          7700          +1.36%
BenchmarkStdlibJSON-4                                                          7176          6914          -3.65%
BenchmarkJSONRPCBuild_Simple_dynamodbPutItem-4                                 16229         12028         -25.89%
BenchmarkJSONUtilBuild_Simple_dynamodbPutItem-4                                15871         11636         -26.68%
BenchmarkEncodingJSONMarshal_Simple_dynamodbPutItem-4                          6447          6393          -0.84%
BenchmarkRESTJSONBuild_Complex_elastictranscoderCreateJobInput-4               272536        213156        -21.79%
BenchmarkRESTBuild_Complex_elastictranscoderCreateJobInput-4                   9023          7757          -14.03%
BenchmarkEncodingJSONMarshal_Complex_elastictranscoderCreateJobInput-4         38727         37823         -2.33%
BenchmarkRESTJSONBuild_Simple_elastictranscoderListJobsByPipeline-4            14645         10682         -27.06%
BenchmarkRESTBuild_Simple_elastictranscoderListJobsByPipeline-4                10505         8934          -14.95%
BenchmarkEncodingJSONMarshal_Simple_elastictranscoderListJobsByPipeline-4      1073          1056          -1.58%
BenchmarkRESTXMLBuild_Complex_cloudfrontCreateDistribution-4                   428704        369790        -13.74%
BenchmarkRESTXMLBuild_Simple_cloudfrontDeleteStreamingDistribution-4           19239         15089         -21.57%
BenchmarkEncodingXMLMarshal_Simple_cloudfrontDeleteStreamingDistribution-4     5547          5447          -1.80%

benchmark                                                                old allocs     new allocs     delta
BenchmarkJSONRPCBuild_Simple_dynamodbPutItem-4                           79             55             -30.38%
BenchmarkJSONUtilBuild_Simple_dynamodbPutItem-4                          77             53             -31.17%
BenchmarkRESTJSONBuild_Complex_elastictranscoderCreateJobInput-4         1131           819            -27.59%
BenchmarkRESTBuild_Complex_elastictranscoderCreateJobInput-4             59             51             -13.56%
BenchmarkRESTJSONBuild_Simple_elastictranscoderListJobsByPipeline-4      80             56             -30.00%
BenchmarkRESTBuild_Simple_elastictranscoderListJobsByPipeline-4          57             49             -14.04%
BenchmarkRESTXMLBuild_Complex_cloudfrontCreateDistribution-4             1832           1584           -13.54%
BenchmarkRESTXMLBuild_Simple_cloudfrontDeleteStreamingDistribution-4     88             64             -27.27%

benchmark                                                                old bytes     new bytes     delta
BenchmarkEC2QueryBuild_Complex_ec2AuthorizeSecurityGroupEgress-4         10922         10920         -0.02%
BenchmarkJSONRPCBuild_Simple_dynamodbPutItem-4                           3873          2624          -32.25%
BenchmarkJSONUtilBuild_Simple_dynamodbPutItem-4                          3809          2560          -32.79%
BenchmarkRESTJSONBuild_Complex_elastictranscoderCreateJobInput-4         41276         25047         -39.32%
BenchmarkRESTBuild_Complex_elastictranscoderCreateJobInput-4             2800          2384          -14.86%
BenchmarkRESTJSONBuild_Simple_elastictranscoderListJobsByPipeline-4      4368          3120          -28.57%
BenchmarkRESTBuild_Simple_elastictranscoderListJobsByPipeline-4          3328          2912          -12.50%
BenchmarkRESTXMLBuild_Complex_cloudfrontCreateDistribution-4             103383        90447         -12.51%
BenchmarkRESTXMLBuild_Simple_cloudfrontDeleteStreamingDistribution-4     9104          7856          -13.71%
```