@nightlyone 2015-11-26
Please put `_ struct{} ` to the front instead of to the end of the struct, see  https://github.com/golang/go/issues/12884 for the reason.