@bpot 2015-11-26
Thanks @nightlyone, I've moved it to the top and verified that it takes no space now.