@jasdel 2015-12-01
Hi @bpot thanks for getting this PR together. I compared the differences locally and using `_` has nearly the same allocs/bytes improvements vs the interface approach. But your method also has a minor improvement on improves on ns/op and is simpler to read I think.

I don't think there are any issues with breaking changes here since metadata<T> was already private so switching it to `_` won't cause an issue.