@jasdel 2015-12-01
Looks good thanks for your time, idea, and submitting the PR.